test: add more-wallets-connect
